Create category and recent languages DDL classes to remove duplication
