+++ app/src/main/java/fr/free/nrw/commons/category/CategoryDao.kt
@@ -4,8 +4,14 @@ import android.annotation.SuppressLint
 import android.content.ContentProviderClient
 import android.content.ContentValues
 import android.database.Cursor
-import android.database.sqlite.SQLiteDatabase
 import android.os.RemoteException
+import fr.free.nrw.commons.category.CategoryTable.ALL_FIELDS
+import fr.free.nrw.commons.category.CategoryTable.COLUMN_DESCRIPTION
+import fr.free.nrw.commons.category.CategoryTable.COLUMN_ID
+import fr.free.nrw.commons.category.CategoryTable.COLUMN_LAST_USED
+import fr.free.nrw.commons.category.CategoryTable.COLUMN_NAME
+import fr.free.nrw.commons.category.CategoryTable.COLUMN_THUMBNAIL
+import fr.free.nrw.commons.category.CategoryTable.COLUMN_TIMES_USED
 
 import java.util.ArrayList
 import java.util.Date
@@ -128,67 +134,6 @@ class CategoryDao @Inject constructor(
             put(COLUMN_THUMBNAIL, category.thumbnail)
             put(COLUMN_LAST_USED, category.lastUsed?.time)
             put(COLUMN_TIMES_USED, category.timesUsed)
-        }
-    }
-
-    companion object Table {
-        const val TABLE_NAME = "categories"
-
-        const val COLUMN_ID = "_id"
-        const val COLUMN_NAME = "name"
-        const val COLUMN_DESCRIPTION = "description"
-        const val COLUMN_THUMBNAIL = "thumbnail"
-        const val COLUMN_LAST_USED = "last_used"
-        const val COLUMN_TIMES_USED = "times_used"
-
-        // NOTE! KEEP IN SAME ORDER AS THEY ARE DEFINED UP THERE. HELPS HARD CODE COLUMN INDICES.
-        val ALL_FIELDS = arrayOf(
-            COLUMN_ID,
-            COLUMN_NAME,
-            COLUMN_DESCRIPTION,
-            COLUMN_THUMBNAIL,
-            COLUMN_LAST_USED,
-            COLUMN_TIMES_USED
-        )
-
-        const val DROP_TABLE_STATEMENT = "DROP TABLE IF EXISTS $TABLE_NAME"
-
-        const val CREATE_TABLE_STATEMENT = "CREATE TABLE $TABLE_NAME (" +
-                "$COLUMN_ID INTEGER PRIMARY KEY," +
-                "$COLUMN_NAME STRING," +
-                "$COLUMN_DESCRIPTION STRING," +
-                "$COLUMN_THUMBNAIL STRING," +
-                "$COLUMN_LAST_USED INTEGER," +
-                "$COLUMN_TIMES_USED INTEGER" +
-                ");"
-
-        @SuppressLint("SQLiteString")
-        fun onCreate(db: SQLiteDatabase) {
-            db.execSQL(CREATE_TABLE_STATEMENT)
-        }
-
-        fun onDelete(db: SQLiteDatabase) {
-            db.execSQL(DROP_TABLE_STATEMENT)
-            onCreate(db)
-        }
-
-        @SuppressLint("SQLiteString")
-        fun onUpdate(db: SQLiteDatabase, from: Int, to: Int) {
-            if (from == to) return
-            if (from < 4) {
-                // doesn't exist yet
-                onUpdate(db, from + 1, to)
-            } else if (from == 4) {
-                // table added in version 5
-                onCreate(db)
-                onUpdate(db, from + 1, to)
-            } else if (from == 5) {
-                onUpdate(db, from + 1, to)
-            } else if (from == 17) {
-                db.execSQL("ALTER TABLE $TABLE_NAME ADD COLUMN description STRING;")
-                db.execSQL("ALTER TABLE $TABLE_NAME ADD COLUMN thumbnail STRING;")
-                onUpdate(db, from + 1, to)
-            }
         }
     }
 }
