+++ app/src/test/kotlin/fr/free/nrw/commons/category/CategoryDaoTest.kt
@@ -17,18 +17,18 @@ import com.nhaarman.mockitokotlin2.mock
 import com.nhaarman.mockitokotlin2.verify
 import com.nhaarman.mockitokotlin2.whenever
 import fr.free.nrw.commons.TestCommonsApplication
-import fr.free.nrw.commons.category.CategoryDao.Table.ALL_FIELDS
-import fr.free.nrw.commons.category.CategoryDao.Table.COLUMN_DESCRIPTION
-import fr.free.nrw.commons.category.CategoryDao.Table.COLUMN_ID
-import fr.free.nrw.commons.category.CategoryDao.Table.COLUMN_LAST_USED
-import fr.free.nrw.commons.category.CategoryDao.Table.COLUMN_NAME
-import fr.free.nrw.commons.category.CategoryDao.Table.COLUMN_THUMBNAIL
-import fr.free.nrw.commons.category.CategoryDao.Table.COLUMN_TIMES_USED
-import fr.free.nrw.commons.category.CategoryDao.Table.CREATE_TABLE_STATEMENT
-import fr.free.nrw.commons.category.CategoryDao.Table.DROP_TABLE_STATEMENT
-import fr.free.nrw.commons.category.CategoryDao.Table.onCreate
-import fr.free.nrw.commons.category.CategoryDao.Table.onDelete
-import fr.free.nrw.commons.category.CategoryDao.Table.onUpdate
+import fr.free.nrw.commons.category.CategoryTable.ALL_FIELDS
+import fr.free.nrw.commons.category.CategoryTable.COLUMN_DESCRIPTION
+import fr.free.nrw.commons.category.CategoryTable.COLUMN_ID
+import fr.free.nrw.commons.category.CategoryTable.COLUMN_LAST_USED
+import fr.free.nrw.commons.category.CategoryTable.COLUMN_NAME
+import fr.free.nrw.commons.category.CategoryTable.COLUMN_THUMBNAIL
+import fr.free.nrw.commons.category.CategoryTable.COLUMN_TIMES_USED
+import fr.free.nrw.commons.category.CategoryTable.CREATE_TABLE_STATEMENT
+import fr.free.nrw.commons.category.CategoryTable.DROP_TABLE_STATEMENT
+import fr.free.nrw.commons.category.CategoryTable.onCreate
+import fr.free.nrw.commons.category.CategoryTable.onDelete
+import fr.free.nrw.commons.category.CategoryTable.onUpdate
 import fr.free.nrw.commons.explore.recentsearches.RecentSearchesContentProvider.Companion.uriForId
 import org.junit.Assert.assertEquals
 import org.junit.Assert.assertNotNull
