+++ app/src/test/kotlin/fr/free/nrw/commons/recentlanguages/RecentLanguagesDaoUnitTest.kt
@@ -12,13 +12,13 @@ import com.nhaarman.mockitokotlin2.inOrder
 import com.nhaarman.mockitokotlin2.mock
 import com.nhaarman.mockitokotlin2.whenever
 import fr.free.nrw.commons.TestCommonsApplication
-import fr.free.nrw.commons.recentlanguages.RecentLanguagesDao.Table.COLUMN_CODE
-import fr.free.nrw.commons.recentlanguages.RecentLanguagesDao.Table.COLUMN_NAME
-import fr.free.nrw.commons.recentlanguages.RecentLanguagesDao.Table.CREATE_TABLE_STATEMENT
-import fr.free.nrw.commons.recentlanguages.RecentLanguagesDao.Table.DROP_TABLE_STATEMENT
-import fr.free.nrw.commons.recentlanguages.RecentLanguagesDao.Table.onCreate
-import fr.free.nrw.commons.recentlanguages.RecentLanguagesDao.Table.onDelete
-import fr.free.nrw.commons.recentlanguages.RecentLanguagesDao.Table.onUpdate
+import fr.free.nrw.commons.recentlanguages.RecentLanguagesTable.COLUMN_CODE
+import fr.free.nrw.commons.recentlanguages.RecentLanguagesTable.COLUMN_NAME
+import fr.free.nrw.commons.recentlanguages.RecentLanguagesTable.CREATE_TABLE_STATEMENT
+import fr.free.nrw.commons.recentlanguages.RecentLanguagesTable.DROP_TABLE_STATEMENT
+import fr.free.nrw.commons.recentlanguages.RecentLanguagesTable.onCreate
+import fr.free.nrw.commons.recentlanguages.RecentLanguagesTable.onDelete
+import fr.free.nrw.commons.recentlanguages.RecentLanguagesTable.onUpdate
 import org.junit.Assert
 import org.junit.Before
 import org.junit.Test
